avoid canonicalization of layout helper read and fixed bug in flat and null-restricted array check snippets

+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/hotspot/replacements/HotSpotReplacementsUtil.java
@@ -373,22 +373,22 @@ public class HotSpotReplacementsUtil {
     }
 
     @Fold
-    public static int flatArrayPattern(@InjectedParameter GraalHotSpotVMConfig config) {
+    public static long flatArrayPattern(@InjectedParameter GraalHotSpotVMConfig config) {
         return config.flatArrayPattern;
     }
 
     @Fold
-    public static int flatArrayMaskInPlace(@InjectedParameter GraalHotSpotVMConfig config) {
+    public static long flatArrayMaskInPlace(@InjectedParameter GraalHotSpotVMConfig config) {
         return config.flatArrayMaskInPlace;
     }
 
     @Fold
-    public static int nullFreeArrayPattern(@InjectedParameter GraalHotSpotVMConfig config) {
+    public static long nullFreeArrayPattern(@InjectedParameter GraalHotSpotVMConfig config) {
         return config.nullFreeArrayPattern;
     }
 
     @Fold
-    public static int nullFreeArrayMaskInPlace(@InjectedParameter GraalHotSpotVMConfig config) {
+    public static long nullFreeArrayMaskInPlace(@InjectedParameter GraalHotSpotVMConfig config) {
         return config.nullFreeArrayMaskInPlace;
     }
